refactor: split headless.py into three private sub-modules (P8-F1)
Extract three cohesive concern groups from the 1,515-line headless.py into
private sub-modules following the _process_*.py → process.py facade pattern:

- _headless_recovery.py: Channel B drain-race recovery, contract nudge, and
  token-merge helpers (~278 lines)
- _headless_path_tokens.py: worktree-path and output-path token extraction,
  validation, and skill-contracts.yaml-derived token set (~114 lines)
- _headless_result.py: SkillResult construction, budget guard, failure
  capture, and session-ID resolution (~541 lines)

headless.py is now a 677-line facade that re-exports all private symbols for
backward-compatible test imports and retains only subprocess orchestration
entry points: _session_log_dir, _resolve_model, _execute_claude_headless,
run_headless_core, and DefaultHeadlessExecutor.

Add tests/arch/test_execution_source_split.py to enforce the split invariants
and size budgets. Update MODULE_CASCADE_EXECUTION in tests/_test_filter.py to
cover the three new private modules.
