+++ tests/_test_filter.py
@@ -165,6 +165,9 @@ MODULE_CASCADE_EXECUTION: dict[str, frozenset[str]] = {
     "anomaly_detection": frozenset({"execution"}),
     "clone_guard": frozenset({"execution"}),
     "_headless_scan": frozenset({"execution"}),
+    "_headless_recovery": frozenset({"execution"}),
+    "_headless_path_tokens": frozenset({"execution"}),
+    "_headless_result": frozenset({"execution"}),
     "_process_io": frozenset({"execution"}),
     "_process_jsonl": frozenset({"execution"}),
     "_process_monitor": frozenset({"execution"}),
